refactor(web): unified sidebar layout for org-settings and preferences
Replace hash-based tabs with nested routes + a secondary sidebar flush
against the global AppSidebar. Split monolithic pages into per-tab files.

- New SettingsLayout (nested SidebarProvider, real shadcn Sidebar primitives)
- Shared SidebarNavLink component (used by NavOrg + SettingsLayout)
- Auto-collapse global sidebar on entry, restore on leave (transient, non-persisted)
- Move p-6 from MainLayout to individual pages (removes -m-6 hack)
- Backwards-compat hash redirects (e.g. /preferences#security)
- Add disableKeyboardShortcut prop to SidebarProvider (avoid Cmd+B conflict)
