+++ apps/web/src/components/sidebar-nav-link.tsx
@@ -1,0 +1,31 @@
+// SPDX-License-Identifier: Apache-2.0
+import type { ReactNode } from "react";
+import { Link } from "react-router-dom";
+import type { LucideIcon } from "lucide-react";
+import { SidebarMenuItem, SidebarMenuButton } from "@/components/ui/sidebar";
+
+export function SidebarNavLink({
+  to,
+  icon: Icon,
+  label,
+  isActive,
+  children,
+}: {
+  to: string;
+  icon: LucideIcon;
+  label: string;
+  isActive: boolean;
+  children?: ReactNode; // badges, indicators
+}) {
+  return (
+    <SidebarMenuItem>
+      <SidebarMenuButton asChild isActive={isActive} tooltip={label}>
+        <Link to={to}>
+          <Icon />
+          <span>{label}</span>
+        </Link>
+      </SidebarMenuButton>
+      {children}
+    </SidebarMenuItem>
+  );
+}
