+++ apps/web/src/pages/tools-page.tsx
@@ -4,5 +4,9 @@ import { ItemTab } from "./item-tab";
 import { toolTabConfig } from "./item-tab-configs";
 
 export function ToolsPage() {
-  return <ItemTab config={toolTabConfig} />;
+  return (
+    <div className="p-6">
+      <ItemTab config={toolTabConfig} />
+    </div>
+  );
 }
